@@ -9,38 +9,36 @@
 		"preview": "vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"format": "prettier --write .",
-		"lint": "prettier --check .",
 		"test:unit": "vitest",
 		"test": "npm run test:unit -- --run"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.9.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.0",
-		"autoprefixer": "^10.4.20",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.20.8",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"autoprefixer": "^10.4.21",
 		"bits-ui": "^0.21.16",
 		"clsx": "^2.1.1",
-		"mode-watcher": "^0.5.0",
-		"prettier": "^3.3.2",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.6.5",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
+		"mode-watcher": "^0.5.1",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"svelte": "^5.28.2",
+		"svelte-check": "^4.1.6",
 		"svelte-sonner": "^0.3.28",
-		"tailwind-merge": "^2.5.5",
-		"tailwind-variants": "^0.3.0",
-		"tailwindcss": "^3.4.9",
-		"typescript": "^5.0.0",
-		"vite": "^6.0.0",
-		"vitest": "^2.0.4"
+		"tailwind-merge": "^2.6.0",
+		"tailwind-variants": "^0.3.1",
+		"tailwindcss": "^3.4.17",
+		"typescript": "^5.8.3",
+		"vite": "^6.3.4",
+		"vitest": "^2.1.9"
 	},
 	"dependencies": {
 		"@tailwindcss/container-queries": "^0.1.1",
-		"@tailwindcss/forms": "^0.5.9",
-		"@tailwindcss/typography": "^0.5.15",
+		"@tailwindcss/forms": "^0.5.10",
+		"@tailwindcss/typography": "^0.5.16",
 		"lucide-svelte": "^0.468.0",
-		"posthog-js": "^1.180.1",
+		"posthog-js": "^1.237.1",
 		"svelte-inview": "^4.0.4",
 		"svelte-motion": "^0.12.2",
 		"svelte-radix": "^2.0.1"
