@@ -1,4 +1,8 @@
+
 <script lang="ts">
+  import { onDestroy, onMount } from 'svelte';
+  import Sparkle from './sparkle.svelte';
+
   interface TSparkle {
     id: string;
     x: string;
@@ -16,56 +20,51 @@
   };
   export let sparklesCount = 5;
 
-  let className = '';
+  export let className: string = '';
   export { className as class };
 
   let sparkles: TSparkle[] = [];
-  import { onMount, onDestroy } from 'svelte';
-  import Sparkle from './sparkle.svelte';
 
-  let generateStar = () => {
-    let starX = `${Math.random() * 100}%`;
-    let starY = `${Math.random() * 100}%`;
-    let color = Math.random() > 0.5 ? colors.first : colors.second;
-    let delay = Math.random() * 4;
-    let scale = Math.random() * 1 + 0.3;
-    let lifespan = Math.random() * 10 + 5;
-    let id = `${starX}-${starY}-${Date.now()}`;
+  const generateStar = (): TSparkle => {
+    const starX = `${Math.random() * 100}%`;
+    const starY = `${Math.random() * 100}%`;
+    const color = Math.random() > 0.5 ? colors.first : colors.second;
+    const delay = Math.random() * 4;
+    const scale = Math.random() * 1 + 0.3;
+    const lifespan = Math.random() * 10 + 5;
+    const id = `${starX}-${starY}-${Date.now()}`;
     return { id, x: starX, y: starY, color, delay, scale, lifespan };
   };
 
-  let initializeStars = async () => {
-    let newSparkles = Array.from({ length: sparklesCount }, generateStar);
-    sparkles = newSparkles;
+  const initializeStars = () => {
+    sparkles = Array.from({ length: sparklesCount }, generateStar);
   };
 
   const updateStars = () => {
-    let temp = sparkles.map((star) => {
-      if (star.lifespan <= 0) {
-        return generateStar();
-      } else {
-        return { ...star, lifespan: star.lifespan - 0.1 };
-      }
-    });
-    sparkles = temp;
+    sparkles = sparkles.map((star) =>
+      star.lifespan <= 0 ? generateStar() : { ...star, lifespan: star.lifespan - 0.1 }
+    );
   };
 
-  let interval: NodeJS.Timer;
-  onMount(async () => {
+  let interval: ReturnType<typeof setInterval>;
+
+  onMount(() => {
     initializeStars();
     interval = setInterval(updateStars, 100);
   });
-  onDestroy(() => clearInterval(interval));
+
+  onDestroy(() => {
+    clearInterval(interval);
+  });
 </script>
 
 <div
-  style:--sparkles-first-color="{colors.first};"
-  style:--sparkles-second-color="{colors.second};"
+  style="--sparkles-first-color: {colors.first}; --sparkles-second-color: {colors.second};"
   class={className}
   {...$$restProps}
 >
   <span class="relative inline-block">
-    {#each sparkles as item, _}
+    {#each sparkles as item}
       <Sparkle {...item} />
     {/each}
     <strong>{text}</strong>
