@@ -1,6 +1,9 @@
 module.exports = {
   extends: ['custom/svelte'],
   rules: {
-    '@typescript-eslint/ban-ts-comment': 'off'
+    '@typescript-eslint/ban-ts-comment': 'off',
+    'no-unused-vars': 'off',
+    'no-console': 'off',
+    '@typescript-eslint/no-unused-vars': 'off'
   }
 };
